requirements.txt
flet>=0.28.0
python-docx>=0.8.11
chromadb>=0.4.0
numpy>=1.24.0
torch>=2.0.0
Pillow>=10.0.0
jieba>=0.42.1
streamlit>=1.28.0
streamlit-chat>=0.1.1
openai>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0